+++ src/app/onbd/page.tsx
@@ -1,0 +1,5 @@
+import OnbdScreen from "@/components/onbd/OnbdScreen";
+
+export default function OnbdPage() {
+  return <OnbdScreen />;
+}